+++ CHANGELOG.md
@@ -189,6 +189,10 @@
 * Enhancements
   * Added `title_prefix` config option.
     `config :brando, title_prefix: "My App >> "`
+  * Brando.Type.Image now has width and height keys. These are set in the `create_image_sizes` function
+
+* Bug fixes
+  * Fixed gifsicle thumbnailing
 
 * Deprecations
   * `model_name` and `model_repr` are now deprecated and removed.